Add unit test for `helm install`
Ref #3143

Also added `exit_on_err` in `bin/test-run`, for DRYing code for failing
fast after tests (we weren't doing that if install_test.go failed during
the non-upgrade).

Note this adds 3 mins of testing in my laptop (out of a total of 13
mins). Would be great if we could consider running some of these in
parallel clusters in the next integration tests infra revamp :wink:

Signed-off-by: Alejandro Pedraza <alejandro@buoyant.io>