Move Code of Conduct from wiki to repo
The Linkerd Community Code of Conduct lives in the wiki:
https://github.com/linkerd/linkerd/wiki/Linkerd-code-of-conduct
Per Github's Community Profile checklist
(https://github.com/linkerd/linkerd2/community), it should live at the
root of the repo.

Copy the contents of the wiki to a markdown file in the root of the
repo. Once merged, we will modify the wiki to point to the repo.

Also update README.md to indicate k8s 1.12+ is required.

Signed-off-by: Andrew Seigner <siggy@buoyant.io>