Fix bad request in top routes tab on empty fields
When clicking on the start button on the top routes tab without
filling the namespace and resource dropdown buttons will cause
a bad request error.

Disabling the `Start` button if the `namespace` and the `resource type`
field of the query object of the state are empty strings will solve
the problem.

Signed-off-by: Pierangelo Di Pilato <pierangelodipilato@gmail.com>
Signed-off-by: pierdipi <pierangelodipilato@gmail.com>