Refactored logic in inject.go's tranform()
Refactored `resourceTransformerInject.inject` to an enum, so we can
clearly distinguish these cases:
- when we want to skip injection altogether (during install because the
helm templates do it for us)
- when we want to inject the proxy directly (using `--manual`)
- when we want to just add the inject annotation

That to be able to more clearly handle each case in the injection
`tranform` function, and fix an issue from #3687 where
`linkerd-heartbeat` was getting injected, after CronJob support was
added.

As a small nice side-effect, the namespace yamls in `cli/cmd/install_test.go` are no
longer needed to be parsed, which explains the yaml properties are no
longer redistributed as shown in the golden files diffs here.